Refactor schools app
