+++ src/apps/schools/api/serializers/__init__.py
@@ -1,0 +1,5 @@
+from .school_serializers import SchoolSerializer
+from .school_manager_serializers import (
+    SchoolManagerViewSerializer,
+    SchoolManagerCreationSerializer,
+)
