+++ src/apps/schools/api/serializers/school_manager_serializers.py
@@ -2,35 +2,9 @@ from rest_framework import serializers
 
 from apps.shared.api.serializers import (
     SchoolSerializerShort,
-    SchoolManagerSerializerShort,
     UserSerializerShort,
 )
-from apps.schools.school_models.schoolModels import School, SchoolManager
-
-
-class SchoolSerializer(serializers.ModelSerializer):
-    """
-    School serializer for view
-        - id
-        - name
-        - address
-        - has_morning_hours
-        - has_afternoon_hours
-        - SchoolManagers (list)
-    """
-
-    school_managers = SchoolManagerSerializerShort(many=True, read_only=True)
-
-    class Meta:
-        model = School
-        fields = (
-            "id",
-            "name",
-            "address",
-            "has_morning_hours",
-            "has_afternoon_hours",
-            "school_managers",
-        )
+from apps.schools.models import SchoolManager
 
 
 class SchoolManagerViewSerializer(serializers.ModelSerializer):
