+++ src/apps/schools/api/views/school_view_set.py
@@ -6,18 +6,15 @@ from apps.shared.api.views import GenericModelViewSet
 from apps.shared.api.permissions import (
     IsAuthenticated,
     IsAdminRole,
-    IsTeacherRole,
     IsSchoolManagerRole,
     OrPermission,
 )
-from apps.schools.school_api.school_serializers.schoolSerializers import (
+from apps.schools.api.serializers import (
     SchoolSerializer,
-    SchoolManagerViewSerializer,
-    SchoolManagerCreationSerializer,
 )
 
 
-class SchoolViewset(GenericModelViewSet):
+class SchoolViewSet(GenericModelViewSet):
     """
     Generic Viewset for School
         - GET: list all schools
@@ -49,25 +46,3 @@ class SchoolViewset(GenericModelViewSet):
         return Response(
             {"detail": "No school found"}, status=status.HTTP_400_BAD_REQUEST
         )
-
-
-class SchoolManagerViewset(GenericModelViewSet):
-    """
-    Generic Viewset for School Manager
-        - GET: list all school managers
-        - POST: create a school manager
-        - GET(id): get a school manager by id
-        - PUT(id): update a school manager by id
-        - DELETE(id): delete a school manager by id
-    """
-
-    serializer_class = SchoolManagerViewSerializer
-    create_serializer_class = SchoolManagerCreationSerializer
-    update_serializer_class = SchoolManagerCreationSerializer
-
-    permission_classes = [IsAuthenticated]
-    list_permission_classes = [IsAdminRole]
-    retrieve_permission_classes = [OrPermission(IsAdminRole, IsSchoolManagerRole)]
-    create_permission_classes = [IsAdminRole]
-    update_permission_classes = [OrPermission(IsAdminRole, IsSchoolManagerRole)]
-    destroy_permission_classes = [IsAdminRole]
