+++ src/apps/schools/models/__init__.py
@@ -1,0 +1,2 @@
+from .school import School
+from .school_manager import SchoolManager
