+++ src/apps/schools/models/school.py
@@ -20,13 +20,3 @@ class School(SharedModelHistorical):
 
     def __str__(self):
         return self.name
-
-
-class SchoolManager(SharedModelHistorical):
-    """
-    Model which represents a school manager
-    """
-
-    # Could be null if the user has not been created yet
-    user = models.OneToOneField(User, on_delete=models.CASCADE, null=True, blank=True)
-    school = models.ForeignKey(School, on_delete=models.CASCADE)
