+++ src/config/urls.py
@@ -24,7 +24,6 @@ schema_view = get_schema_view(
 # Importing apps urls
 from apps.requests.request_api.request_routers import urlpatterns as requests_urls
 from apps.homeworks.homework_api.homework_routers import urlpatterns as homeworks_urls
-from apps.schools.school_api.school_routers import urlpatterns as schools_urls
 
 
 # API urls
@@ -47,7 +46,7 @@ urlpatterns = [
     path("api/auth/", include("apps.authentication.api.urls")),
     path("api/users/", include("apps.users.api.urls")),
     path("api/students/", include("apps.students.api.urls")),
-    path("api/schools/", include(schools_urls)),
+    path("api/schools/", include("apps.schools.api.urls")),
     path("api/subjects/", include("apps.subjects.api.urls")),
     path("api/teachers/", include("apps.teachers.api.urls")),
     path("api/requests/", include(requests_urls)),
